requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.3.0
tensorflow==2.15.0